modular activity creation page
