+++ picto-app/src/pages/ActivityCreation/ActivityOptionsCard.tsx
@@ -1,0 +1,103 @@
+import React from "react";
+import { ClockIcon } from "@heroicons/react/24/outline";
+import { FormControlLabel } from "@mui/material";
+import { ActivityIstance } from "@/utils/Types";
+import IOSSwitch from "../PagesUiComponents/IOSSwitch";
+
+interface Props {
+    formValues: ActivityIstance;
+    setFormValues: React.Dispatch<React.SetStateAction<ActivityIstance>>;
+}
+
+const ActivityOptionsCard: React.FC<Props> = ({ formValues, setFormValues }) => {
+    const handleCheckedAddChrono = (enabled: boolean) => {
+        setFormValues({ ...formValues, chrono: { ...formValues.chrono, isEnabled: enabled } });
+    };
+
+    const onChangeChronoTime = (e: React.ChangeEvent<HTMLInputElement>) => {
+        const value = e.target.value;
+        if (value === "") {
+            setFormValues({ ...formValues, chrono: { ...formValues.chrono, [e.target.name]: 0 } });
+        } else if (e.target.name && Number(value) >= 60) {
+            const newMinutes = formValues.chrono.minutes + Math.floor(Number(value) / 60);
+            setFormValues({ ...formValues, chrono: { ...formValues.chrono, minutes: newMinutes, seconds: Number(value) % 60 } });
+        } else {
+            setFormValues({ ...formValues, chrono: { ...formValues.chrono, [e.target.name]: Number(value) } });
+        }
+    };
+
+    return (
+        <div className="options-card">
+            <div className="card-title-group">
+                <h3 className="main-left_title">Options</h3>
+                <p className="card-subtitle">Paramètres supplémentaires pour l'activité</p>
+            </div>
+
+            <div className="options-row-list">
+                <div className="option-section">
+                    <div className="option-row">
+                        <FormControlLabel
+                            control={
+                                <IOSSwitch
+                                    onChange={(_, checked) => setFormValues({ ...formValues, authoriseEdit: checked })}
+                                    checked={formValues.authoriseEdit}
+                                />
+                            }
+                            label=""
+                        />
+                        <div className="option-row__text">
+                            <span className="option-row__label-title">Autoriser l'édition</span>
+                            <span className="option-row__label-desc">Les participants peuvent modifier leurs réponses</span>
+                        </div>
+                    </div>
+                </div>
+
+                <div className="option-section">
+                    <div className="option-row">
+                        <FormControlLabel
+                            control={
+                                <IOSSwitch
+                                    onChange={(_, checked) => handleCheckedAddChrono(checked)}
+                                    checked={formValues.chrono.isEnabled}
+                                />
+                            }
+                            label=""
+                        />
+                        <div className="option-row__text">
+                            <span className="option-row__label-title">Chronomètre</span>
+                            <span className="option-row__label-desc">Définir une durée limite pour l'activité</span>
+                        </div>
+                    </div>
+                    {formValues.chrono.isEnabled && (
+                        <div className="add_clock-field">
+                            <ClockIcon width={15} height={15} className="chrono-icon" />
+                            <div className="chrono-inputs">
+                                <div className="chrono-field-group">
+                                    <input type="number" name="minutes"
+                                        value={formValues.chrono.minutes}
+                                        min={0} max={60}
+                                        onChange={onChangeChronoTime}
+                                        className="chrono_field"
+                                        placeholder="00" />
+                                    <span className="chrono-label">min</span>
+                                </div>
+                                <span className="chrono-separator">:</span>
+                                <div className="chrono-field-group">
+                                    <input type="number" name="seconds"
+                                        value={formValues.chrono.seconds}
+                                        min={0} max={60}
+                                        onChange={onChangeChronoTime}
+                                        className="chrono_field"
+                                        placeholder="00" />
+                                    <span className="chrono-label">sec</span>
+                                </div>
+                            </div>
+                        </div>
+                    )}
+                </div>
+            </div>
+        </div>
+    );
+};
+
+export default React.memo(ActivityOptionsCard);
